requirements.txt
django
djangorestframework
psycopg2-binary
redis
django-redis
requests
python-dotenv
dj_database_url